webhook/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
PyJWT==2.10.1
aiohttp==3.11.11
python-multipart==0.0.20
